Fixed changes to .ts file to test on front-end
