+++ ClientApp/src/app/overlay-panel/overlay-panel.component.html
@@ -1,0 +1,9 @@
+<p-overlaypanel
+    [dismissable]="true"
+    [showCloseIcon]="true"
+    [appendTo]="body"
+    [style]="{width: '300px'}"
+    #overlayPanel
+></p-overlaypanel>
+
+<p-toast></p-toast>
