--- ClientApp/src/app/overlay-panel/overlay-panel.html
@@ -1,24 +1,0 @@
-<p-button [label]="selectedProduct ? selectedProduct.name : 'Select a Product'" icon="pi pi-search" (click)="op.toggle($event)"></p-button>
-
-<p-overlayPanel #op [showCloseIcon]="true" [style]="{width: '450px'}">
-    <ng-template pTemplate>
-        <p-table [value]="products" selectionMode="single" [(selection)]="selectedProduct" (onRowSelect)="onRowSelect($event)" [paginator]="true" [rows]="5" responsiveLayout="scroll">
-            <ng-template pTemplate="header">
-                <tr>
-                    <th pSortableColumn="name">Name<p-sortIcon field="name"></p-sortIcon></th>
-                    <th>Image</th>
-                    <th pSortableColumn="price">Price <p-sortIcon field="price"></p-sortIcon></th>
-                </tr>
-            </ng-template>
-            <ng-template pTemplate="body" let-rowData let-product>
-                <tr [pSelectableRow]="rowData">
-                    <td>{{product.name}}</td>
-                    <td><img src="https://primefaces.org/cdn/primeng/images/demo/product/{{product.image}}" [alt]="product.image" class="product-image" /></td>
-                    <td>{{product.price}}</td>
-                </tr>
-            </ng-template>
-        </p-table>
-    </ng-template>
-</p-overlayPanel>
-
-<p-toast></p-toast>
